backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
motor==3.6.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
dnspython==2.6.1
